Bug 1962497 - Fix RELRHACK_LINKER setting when linker name is target triple prefixed r=glandium
RELRHACK_LINKER is used as the name of a binary installed in a
directory specified with -B to override the linker. Both Clang and
GCC will only look for a binary named "ld" (or "ld.$fuse_ld_setting"
if -fuse-ld= is specified) in the -B directories, which means that
if the linker name does not follow this pattern, for example if it
is named $target_triple-ld", the relrhack linker will not be found,
the compiler will use the normal linker and the link will fail. To fix
this problem, use the correct pattern to name the relrhack executable.

Differential Revision: https://phabricator.services.mozilla.com/D246545

UltraBlame original commit: 7d714353f8a85796980d378b2b51a8c807d6aca2