Bug 1964863 - Prevent child SharedDataMaps in parent and vice versa r=nimbus-reviewers,relud
Previously it was possible to create parent and child `SharedDataMap`s in
the parent or child process, but only so that they could be tested in
xpcshell tests. The xpcshell test has been rewritten as a browser test
so that it can do real cross-process communication, so that is no longer
required.

he `_shutdownBlocker` member is now initialized in parent
`SharedDataMap`s because the the child does not need to block shutdown
and this property was never accessed in the child process.

Some cleanups have been made to `browser_experimentapi_child.js` to not
re-import `Assert.sys.mjs` (and therefore override the built-in `Assert`
that reports errors to the parent correctly) and to use
`SpecialPowers.addTaskImport` to add imports into the content task
scope.

Differential Revision: https://phabricator.services.mozilla.com/D248549

UltraBlame original commit: 8dc16469751cad77cd3686748a6471275ec1a087