Bug 1963549 - Add profile migration to remove outdated sidebar attrs from XULStore. r=Gijs,sidebar-reviewers,firefox-desktop-core-reviewers ,sessionstore-reviewers,sthompson
There was some discussion on whether the XULStore data should be written back into Session Store or Prefs. My concern there is that since we no longer *write* to XULStore, there is a good chance that this data is outdated anyways, and we'd be overwriting data in other places that *are* up-to-date. My vote is to simply discard this data, it seems that users were already comfortable with clearing their xulstore.json, so it doesn't seem like an issue to do the same programmatically.

Differential Revision: https://phabricator.services.mozilla.com/D248476

UltraBlame original commit: 35212b487fd942d910254b12ba888fc35f746159