Bug 1965595: Change HTMLImageElement::NaturalSize return type to CSSIntSize. r=emilio
Before this patch, the return-type was nsIntSize, which is ambiguous about
units.

I'm switching to CSSIntSize because I believe the units are CSS Pixels. Note
that this function is essentially implementing the spec text at [1] which
includes an example that explicitly describes the result as "CSS Pixels". Also,
this function is analogous to nsImageLoadingContent::GetWidthHeightForImage()
which also returns CSSIntSize.  (And it's convenient to have those share the
same return type, because I'm planning to have these functions share more of
their internals in a later patch.)

[1] https://html.spec.whatwg.org/multipage/images.html#density-corrected-intrinsic-width-and-height

Differential Revision: https://phabricator.services.mozilla.com/D248674

UltraBlame original commit: 72ee50ceade744f48d24107ff9e5acacfafba90b