Bug 1325145 - [devtools] Wait for setProperty/preview to be settled before updating TextPropertyEditor on value inplace editor destroy. r=devtools-reviewers,bomsy.
This patch updates InplaceEditor so it passes the result of the options.done callback
to the options.destroy one, so we can orchestrate things in callsites.

For the declaration value, this allows us to properly wait for the value to be
previewed/set before updating the view, so all underlying data are up to date.

A test case is added to check that this fixes the original issue (warning icon
being displayed on declaration after clearing the value but dismissing the edit).
We also add an option to the `openRuleView` helper to avoid overriding the `debounce`
method, so test can replicate more accurately what actually happens.

Differential Revision: https://phabricator.services.mozilla.com/D247225

UltraBlame original commit: 708aeef0da4191c16b15f3724c5ec8b8a87df9bb