Bug 1467846: Part 6: Remove Live list from OffThreadPromiseRuntimeState; r=arai
This patch removes the live state all together. The reasoning for this is as follows:

Previously, the live set had the following responsibilities:
- the set was a set of raw pointers. This functioned as a list of tasks that would potentially be
  dispatched and had been initialized.
- If we attempted to dispatch, but failed to do so (such as during shutdown), we would increment the
  numFailed_ (previously termed numCanceled_) counter. However with the shift to using a UniquePtr
  in the embedding side, we now transfer ownership of the UniquePtr when a task fails to the dead()
  list. We no longer need the live list to manage this information, instead we can iterate over the
  dead Fifo.
- In the future, the plan was the use the live list for cancellable. However, this too is better
  represented as an owned list, as the runtime can hold the token for cancellable tasks until it is
  dispatched. This becomes clearer when we implement the WaitNotifyTask, see further in the patch queue.

Now, the live set has been replaced with a numRegistered_ counter. This is the number of tasks
currently registered and not yet completed. If a task fails, or is registered as cancellable, it
will go into one of two collections for tracking owned Dispatchables that potentially need to be
cleaned up on shutdown.

In the case of Cancellables, they are blocked from using the regular DispatchResolveAndDestroy
method. They now have a special method: releaseCancellableAndDispatch, that will remove itself from
the cancellable list, and transfer ownership of the UniquePtr to the embedding. If this fails, it
will follow the same failure route as for all other dispatchables.

In the case of dead Dispatchables: They can not be redispatched, instead they wait until shutdown to
be cleared. They are cleared in a very similar way as the old live list, but with less book-keeping.

We still cannot remove the js_delete pattern, as WASM relies on it.

Depends on D244936

Differential Revision: https://phabricator.services.mozilla.com/D244937

UltraBlame original commit: 4d05b5882f15a5049cb92b116181892f46151631