Bug 1967025: Add a "layout code reviewer checklist" doc, to gather best practices for code review in layout code. r=TYLin,tlouw
As noted in the doc, this is meant to extend the general checklist at
https://firefox-source-docs.mozilla.org/contributing/reviewer_checklist.html
with examples and special cases that are particular to working in layout code.

The list is short for now; this is just a start.

Differential Revision: https://phabricator.services.mozilla.com/D249859

UltraBlame original commit: e27f307551a8ff9e0d79d20bfdf5131ecdef5d78