Bug 1965348 - Add `generate-python-lockfiles` `mach` command r=ahal
This command takes the <site>.txt files at `python/sites/` and converts them to
pyproject.toml files that `uv` can use to make lockfiles.

The `mach` site is locked by itself, then all other sites are locked with their own
dependencies, plus the dependencies of the `mach` site (since they're inherited on
site activation, just specified in different files).

The generated files are not intended to be committed to the repo, because we only care if
we can lock, we don't use them for anything else. If we can lock, the site is valid and won't
have compatibility errors.

A `--keep-lockfiles` arg was added that prevents the temporary directory containing the generated
files from being deleted, to help with troubleshooting when a site is not able to be locked.

This command can be run by itself if updating the a <site>.txt file manually with a `pypi:` dependency, and
it will be integrated with `./mach vendor python` and `test_site_compatibility.py` in the near future.

Differential Revision: https://phabricator.services.mozilla.com/D249492

UltraBlame original commit: 3ff59cbd95b9702836e9e834d004198c2491aae9