Bug 1820252 - Add TranslationsDocument test helpers r=translations-reviewers,gregtatum
This patch adds a few public functsion to the TranslationsDocument
as well as the TranslationsChild actor that are primarily intended
for use during testing. These functions help all of our tests not
only ensure that the content is translated correctly, but also that
the TranslationsDocument is upholding all of its invariants.

Differential Revision: https://phabricator.services.mozilla.com/D249985

UltraBlame original commit: c9227482c87a8db895b27328fbd160cdc70e34da