Bug 1581963 - [devtools] Implement a new remote function to gather the profile data, using the bulk api r=devtools-backward-compat-reviewers,profiler-reviewers,ochameau,mstange
The devtools protocol's bulk API is much more efficient, as we're
transfering the protocol data in a binary form instead of serializing it
using the devtools protocol.
In addition to that, the profile data is now gzipped before sending it.

Differential Revision: https://phabricator.services.mozilla.com/D244341

UltraBlame original commit: 18010dd7c6a386f1ed4c02750968204f5d76516b