Bug 1327973 - [devtools] Remove count/score sort in markup/style editor autocomplete suggestions. r=devtools-reviewers,jdescottes.
The suggestions were sorted by the number of element matching the item, which could
be confusing and get in the way of the user when they wanted to search for the
least used version.
We now sort the suggestions by "type" (id, class, tag, …), and then, within those
groups, alphabetically.

Differential Revision: https://phabricator.services.mozilla.com/D249768

UltraBlame original commit: 410248095e235e7fd3fd8d33be38746210bfbe39