Bug 1957557, part 7 - Create Text Fragments: Compute word boundary distances for context terms. r=dom-core,farre
The distances are always in search direction:
For prefix and end terms, the positions are
relative to the end of the term and use the word
start boundary.
For start and suffix, they are relative to its
beginning and use the word end boundary.

Differential Revision: https://phabricator.services.mozilla.com/D246194

UltraBlame original commit: 29735d438f2634e2c6c1d8e80857bfb144d36483