Bug 1957557, part 8 - Create Text Fragments: Find all matches for the target range in the partial document. r=dom-core,farre
Search the partial document to find all
other matches for the content of the target range.
Then, for these matches, compute and save the
common lengths of the context terms.

For range-based matching, this needs to be done
separately for the start term and the end term.

Differential Revision: https://phabricator.services.mozilla.com/D246195

UltraBlame original commit: 1ae864cdf71c71ccedafa75a70420187b19b3296