Bug 1968431 - Introduce infrastructure to re-cascade implicitly inherited zoom-dependent properties if needed. r=jfkthame
This is infrastructure for a more generic version of bug 1911195. It
doesn't fix everything because:

 * We still have an if is_unset { return; } in cascade.rs which we need
   to deal with. That means that we don't deal with explicit `inherit`.
   Note that was the case with the other bug afaict.

 * It still doesn't automatically derive the length dependency, relying
   on a list in data.py.

The general approach is basically taking the inherited computed value
from the parent, but instead of just copying it blindly we re-compute it
with the specified zoom. This is handled only in cascade.rs, via
overriding the effective_zoom (arguably a bit hacky, but does the
trick).

I think in general it's a much cleaner approach, and fairly easy to
extend to explicit inheritance and non-inherited properties (for things
like e.g. width: inherit we need to do the same).

Differential Revision: https://phabricator.services.mozilla.com/D251121

UltraBlame original commit: d233e5e360ab3fde8c9f2f58d070df61b085218f