Bug 1943467 - wasm: Implement Wasm ABI in ABIArgGenerator using mode instead of inheritance. r=yury
WasmABIArgGenerator worked by inheriting ABIGenerator and then changing the stack
offset to account for the size of the instance slots. masm only ever stored
an ABIArgGenerator on itself for masm.callWithABI, but would assign a
WasmABIArgGenerator onto itself in masm.setupABICallHelper. This is technically
broken and wouldn't work if a field was added to WasmABIArgGenerator.

This commit refactors this so that there is no WasmABIArgGenerator, but
instead ABIArgGenerator takes a mode flag. No functional changes.

This also allows us to tighten up the assertions in the platform specific
arg generators for special cases they have for the wasm ABI that should
not be hit when calling system ABI functions.

Differential Revision: https://phabricator.services.mozilla.com/D244434

UltraBlame original commit: 6190e0ac80712713c7ae051756025bd863a2ecc4