Bug 1964973: Remember the TURN protocol (UDP, TCP) in relay candidates. r=mjf
We had been counting on this information being preserved in the candidate's
base address, but for allocated relay candidates that is the TURN server's
address (which will be UDP), not ours. This was causing an incorrect priority
in ICE checks.

Depends on D250490

Differential Revision: https://phabricator.services.mozilla.com/D250491

UltraBlame original commit: d4dad9cccd10b3ff63cfb3cbd9a263db075b9849