Bug 1968925 - Better themed buttons in GTK. r=stransky,desktop-theme-reviewers,dao
This patch got a bit larger than expected mostly to deal well with
windows HCM, but the TLDR is that we introduce Button*border colors, and
deal with them in the non-native theme, rather than hard-coding them.

We also leave the previous hardcoded colors in the standin set, so that
web content doesn't change behavior, and reorder a bit the colors to be
closer together.

Differential Revision: https://phabricator.services.mozilla.com/D251648

UltraBlame original commit: dc2493908a75b28c986de2872768403658da62b9