Bug 1682759. Support exif data in pngs. r=gfx-reviewers,nical
Since the reorienting surface pipe doesn't support some features we need to prioritize. If the png has animation then we ignore exif data. We disable progressive display and interlacing if the png has exif data and the effect is not the identity (we don't take advantage of the improved experience while the image is loading, but it still works fine when fully loaded).

Differential Revision: https://phabricator.services.mozilla.com/D252014

UltraBlame original commit: 99942e19563aedddadfccb8d326011ed6cdcfc8d