Bug 1969853. Fix css/css-images/image-orientation/image-orientation-exif-png-3.html. r=layout-reviewers,emilio
Refering to

https://wpt.fyi/results/css/css-images/image-orientation/image-orientation-exif-png-3.html?label=experimental&label=master&aligned

one can see that this test fails on all browsers for the same reason: the second image is slightly higher than in the reference. The reason is that the reference uses img elements which are inline replaced elements by default, and the test uses divs, which are block by default. So we make the divs inline-block in the test so the images line up.

(Note Firefox currently fails this test for a second reason: it doesn't support exif in png, but bug 1682759 will land and resolve that shortly.)

Differential Revision: https://phabricator.services.mozilla.com/D252120

UltraBlame original commit: 2309b310c9f85d47f7968be67d57510ff6da3ecf