Bug 1970005 - Experimenter warning to use fxms-message-12 thru 15 only for Fx140+ targeting r=omc-reviewers,jprickett
Bug 1965869 was fixed in 140, meaning that experiments on fxms-message-12 through fxms-message-15 targeting users using versions in 139 and earlier will likely fail most of the time.

Explicitly make this clear in the "description" field of FeatureManifest.yaml, which is displayed in the Experiment UI when one is selecting a feature ID.  This will be deployed to Experimenter shortly after landing in mozilla-central.

Differential Revision: https://phabricator.services.mozilla.com/D252258

UltraBlame original commit: 9a27f9ad7eb946d8d30c81747c77e277f90acb61