Bug 1859474 - Don't release TCPSocket on socket thread. r=necko-reviewers,kershaw
Since TCPSocket might be released on the socket thread, it might cause a
crash when TCPSocket is destroyed on this thread.

Since we use mozilla::SyncRunnable::DispatchToThread, this is synchronous
call. So it is unnecessary to touch reference count on this thread.

Differential Revision: https://phabricator.services.mozilla.com/D251990

UltraBlame original commit: 32b1a5a7c6f5f09d5694ed47af13969f69d2c70b