Bug 1960433: Load jitcode for trial inlining infallibly r=jandem
When we passed the ICScript in the JSContext, we had to be certain that we were calling baseline code. If we instead called a different tier, it would leave the ICScript in the context, and a subsequent call to baseline code might pick it up, leading to a mismatched set of ICs. Now that we're passing it in the stub frame, it is always safe to call a lower tier, so we can simplify this code.

Differential Revision: https://phabricator.services.mozilla.com/D251072

UltraBlame original commit: 7d304c4a18ba5e56e7a9071a05264138e4081d5f