Bug 1969912 [wpt PR 52901] - (Partially) fix paint-timing IDL harness test.,
Automatic update from web-platform-tests
(Partially) fix paint-timing IDL harness test.

This test currently fails with an error regarding DomHighResTimeStamp
missing, and that's because the test, idlharness.window.js, doesn't
declare a dependency on the 'hr-time'. Fixing that in this cl.

However, this doesn't actually make the test pass, because the spec for
the performance timeline is broken: 1) It requires that there be an id
attribute on every performance entry, of type uint64. This is
impossible, because LargestContentFulPaint entries have also an id
attribute, and it's a string. So this cannot work (and that's the error
now in the expected file). 2) There's also the (still masked)
navigationId error, another attribute of PerformanceEntry. This is
spec'd as a uint64, in Chrome it's currently a string. I'll try to fix
that one.

Just to demonstrate the two points above, I'll upload a first patchset
with this change, which shows how to make the idlharness.window.js test
actually pass, but this requires commenting out a couple lines from the
spec IDL, which I can't do in this (Chromium) change. So then I'll undo
that, in patchset 2, and just fix the test with a new expected output.

Should you later want to see these two (Chromium) patchsets, you may
find this change at crrev.com/c/6604075

Bug: 420166099
Change-Id: Iab6a6185f77853e2380a4e11fe235e8ab61bcff5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6604075
Commit-Queue: Johannes Henkel <johanneschromium.org>
Reviewed-by: Noam Rosenthal <nrosenthalgoogle.com>
Cr-Commit-Position: refs/heads/main{#1468173}

--

wpt-commits: 2d54aa0f4d20f5b4d4aab5956ccb1adb9de280cd
wpt-pr: 52901

Differential Revision: https://phabricator.services.mozilla.com/D252821

UltraBlame original commit: c5f8eeaadc53a8432ab0ff48fa58325c6e362ce4