Bug 1781153 - Simplify use of AXTitle/AXDescription. r=morgan
Generally speaking, only aria-label names should be exposed as AXDescription.

There are a few exceptions, so we still need to distinguish this by role. I refactored
the logic a bit so that AXTitle and AXDescriptions are simpler to read and offer the exact
inverse of each other.

Differential Revision: https://phabricator.services.mozilla.com/D252415

UltraBlame original commit: 530dd27622803bc3636be9e3ca2749e927608c21