Bug 1969164 - close window when dragging tab group out of window r=jswinarton,tabbrowser-reviewers,dao
When a window consists only of a tab group and its tabs, dragging the tab group into the tab strip of another window can leave the starting window in a broken, empty state.

The normal behavior when dragging a tab group is to temporarily collapse the tab group, but this seems to have an unintended interaction with the Tabbrowser logic that decides whether to close a window when the last tab is removed. At the time of the last tab being removed, it is in a temporarily collapsed group and therefore not visible.

This patch hardens the logic for calculating whether a tab is the last tab in a window. It ensures that the window will close if the last tab in the window is in a collapsed tab group.

Differential Revision: https://phabricator.services.mozilla.com/D251943

UltraBlame original commit: 0f12899d99ae7153ed0c085e794e854ee6105c24