Bug 1969962 Part 3 - Share grid item intrinsic size min-size clamping between spanning and non-spanning code r=dholbert,tlouw
This is a part of the work to fully share the intrinsic sizing code, as this
code is equivalent but written differently for the two uses.

Ideally the non-spanning code that uses ShouldApplyAutoMinSize will be merged
with the spanning code that checks eContentBasedAutoMinSize, but that will
require further effort.

Differential Revision: https://phabricator.services.mozilla.com/D252212

UltraBlame original commit: 90e188336ec0ca12c88f44ff6250d02e0e13599d