Bug 1964949 - Implement the selector matching for <pt-name-and-class-selector>. r=emilio,view-transitions-reviewers
We added a special case for matching `<pt-name-and-class-selector>` because
we have to check its class list in the related captured element.

This patch also added a class list in captured element, and updated it
when capturing old state and new state. However, the name is a
tree-scoped name, so we fail a test. Let's leave it to Bug 1970954.

Also, there are a lot of intermittents on Android (note: their passing rates
are about 70%-90%). I checked them with adb logcat (running on the Android
emulator) and their reftest results in try. In these failed tests, we didn't
render anything, but the selector matching works as I expected.

Besides, without this patch, some of the tests are intermittents already
(especially for the tests which are passed already in wpt.fyi,
e.g. xxx-mismatch-xxx). That means they may not be related to this patch,
and these intermittents are probably belonging to some of the existing bugs,
especially on Android. Perhaps we have some race conditions on Android when
rendering view transition pseudo-elements for some cases, or the tests are
not stable enough.

Differential Revision: https://phabricator.services.mozilla.com/D252903

UltraBlame original commit: ca3bdda9fbfd2aaecb0ec014128bdce7d49f117b