Bug 1971450 - Remove internal appearance values for meter / progress / range thumb / chunk. r=dshin
This is technically a behavior change because we don't really look at
the appearance value of the progressbar etc anymore. We could do that if
wanted, but it's more consistent to just look at the individual frame
(looking forward to appearance: base and other potential compat
improvements we could make in the area).

Differential Revision: https://phabricator.services.mozilla.com/D253256

UltraBlame original commit: 13735d3c75cb7a6fabbdcae93bdfce3376d526e4