Bug 1972914 - Restore border colors on Windows HCM. r=win-reviewers,handyman
Bug 1968925 moved the special-case to just use system colors, and I made
some mistakes on the HCM colors.

Factor all the HCM special cases to a single block, so that it's easier
to reason about and audit.

This restores the behavior of form controls on HCM that we had
pre-regression and have on release.

Differential Revision: https://phabricator.services.mozilla.com/D254354

UltraBlame original commit: e7d4bc1552286ff20b8469665fc14f8423e769fb