Bug 1971487 - Stop setting lwtheme attribute in contentTheme.js, and revert the previous patch. r=desktop-theme-reviewers,dao
This is to land after the soft freeze.

I added this for bug 1877298, and since then we stopped trying to do
what that bug was doing.

Now it's unused, and after the regressing bug it causes problems,
because the presence of a theme is not indicative of whether the page
should itself be themed (that's for [lwt-{sidebar,newtab}] attributes to
decide).

Differential Revision: https://phabricator.services.mozilla.com/D254528

UltraBlame original commit: b9cb5d3dd933496ae5b2cb927d162e8b1e67359c