Bug 1968122 - Update tier 2 CTS expectations. r=webgpu-reviewers,nical
I had to demote 2 tests because they are testing promise ordering that
is not guaranteed by the specification. I will open an issue on the
spec repo to clarify the situation.

The new timeouts are actually existing timeouts that were not surfaced
by the harness. I've looked at a recent CTS run without the patch stack
and the tests time out but the harness reports TEST-OK. Locally,
the tests also run at the same speed with and without the patch stack.

I'm not sure what change in the patch stack caused the harness to now
report the timeout but I think this is now working as intended.

Differential Revision: https://phabricator.services.mozilla.com/D254255

UltraBlame original commit: 12619b90dd5e377491c1662432bce3612906c65b