Bug 1972116 - wasm loop unroller: correctly handle dependences through memory.  r=iain.
When duplicating MIR nodes in the original loop body, the load dependency field
was not being updated.

It is necessary for the load dependency field of a duplicated node to be
updated to point at the corresponding duplicated store node in the same
iteration-copy.

The result was that the load dependency fields ended up still pointing at the
store node in the original loop, which eventually becomes the peeled iteration,
outside the loop.  Post-unrolling GVN could then common up all the loads and
replace them with the value loaded in the peeled iteration, so that the loaded
value was, incorrectly, the value stored by the first iteration of the
(original) loop.

This patch fixes it in a straightforward way, and adds a regression test.

Differential Revision: https://phabricator.services.mozilla.com/D254220

UltraBlame original commit: 9e20e0a11b4c8a0ecae3aac88abd4ed4cc1351fb