Bug 1973418 - Fix promise crash with remote encoders that are synchronously created. r=media-playback-reviewers,padenot
If we call RemoteMediaDataEncoderChild's Init method because Construct
returns, we defer it until Construct finishes. Because of how
InvokeAsync works, we ended up chaining on the same promise twice, which
violates the contract for exclusive promises. This patch removes the
chaining on the Construct pathway to avoid this.

Differential Revision: https://phabricator.services.mozilla.com/D254624

UltraBlame original commit: dfad54d09fdad8c6f0800492bc28a2405eb92da7