Bug 1962870 - part 1: Reimplement `EventStateManager::MaybeFirePointerCancel` r=smaug,edgar,dom-core
`EventStateManager::MaybeFirePointerCancel()` was designed to dispatch
`ePointerCancel` to current event target.  However, this design is not
reasonable.

First, the method is used only by
`EventStateManager::GenerateDragGesture()` and the spec defines that
the `pointercancel` event target should be the preceding `dragstart`
event target [1].  So, specifying the target in the caller side must be
better for the complicated cases.

Next, `mCurrentTarget` may be a text frame under the cursor.  Therefore,
the method may dispatch to the text node rather than its element.

Next, `mCurrentTarget` may be set to `nullptr` if the frame is reframed
by a `dragstart` event listener.  Then, the method does not dispatch
`ePointerCancel` to the DOM.

Additionally, when `pointercancel` is dispatched, the steps of the
"fire a pointer event" should be handled.  Then, dispatching
`pointercancel` should process pending pointer capture state before.
However, this is not handled.

The definition of "fire a pointer event" is complicated and maybe that
could be updated by further spec changes.  Therefore, I believe that
the steps should be encapsulated into `PointerEventHandler`.  Thus,
this patch adds `PointerEventHandler::DispatchPointerEventWithTarget()`
overloads which skip considering the event target with a hit test.

Finally, `EventStateManager::MaybeDispatchPointerCancel()` which is a
new method replacing `EventStateManager::MaybeFirePointerCancel()` takes
the target, adjust the target to an `Element` and use
`PointerEventHandler`.

Note that, as far as I've tested, if the `dragstart` target is removed
from the DOM, Chrome retargets the following `pointercancel` to the
closest ancestor which is still connected to the DOM.  This tracking
is required 2 places, when dispatching `eDragStart` and
`ePointerGotCapture`/`ePointerLostCapture`.  Therefore, I added a
general helper class, `AutoConnectedAncestorTracker` which tracks the
closest connected ancestor within `PresShell::ContentWillBeRemoved`.

1. https://html.spec.whatwg.org/#drag-and-drop-processing-model

Differential Revision: https://phabricator.services.mozilla.com/D253622

UltraBlame original commit: 943efbe6491ccc9271d183d1cc1259f17441ddab