Bug 1972248 - Try all available encoders in encoder creation r=media-playback-reviewers,jolin
This patch tries to create the encoder from a sorted list instead of
only trying to create the top-choice encoder from the list. This makes
it possible for us to skip some encoders that are not compatible with
the given configuration.

Differential Revision: https://phabricator.services.mozilla.com/D253761

UltraBlame original commit: d3c70dd220acaa573eed9889579b41a624599961